+++ src/pages/instructorRegister/components/Description/index.css.ts
