+++ src/pages/instructorRegister/components/InstructorRegisterFunnel/ImageUploadStep/index.css.ts
