+++ src/pages/instructorRegister/components/InstructorRegisterFunnel/IntroductionStep/index.css.ts
