+++ src/pages/instructorRegister/components/InstructorRegisterFunnel/VideoLinkStep/index.css.ts
