+++ src/pages/instructorRegister/components/InstructorRegisterFunnel/index.css.ts
